requirements.txt
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cryptography==44.0.0
deep-translator==1.11.4
idna==3.10
packaging==24.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
pycparser==2.22
PyMuPDF==1.25.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
reportlab==4.2.5
requests==2.32.3
soupsieve==2.6
typing==3.7.4.3
urllib3==2.2.3
